add function some_files
