+++ product/pro_add_done.php
@@ -30,12 +30,13 @@ else
 try
 {
 
-$pro_name = $_POST['name'];
-$pro_price = $_POST['price'];
-$pro_image_name = $_POST['image_name']; // 画像のファイル名
+require_once('../common/common.php');
 
-$pro_name = htmlspecialchars($pro_name, ENT_QUOTES, 'UTF-8');
-$pro_price = htmlspecialchars($pro_price, ENT_QUOTES, 'UTF-8');
+$post = sanitize($_POST);
+
+$pro_name = $post['name'];
+$pro_price = $post['price'];
+$pro_image_name = $post['image_name']; // 画像のファイル名
 
 $dsn = 'mysql:dbname=shop;host=localhost;charset=utf8';
 $user = 'root';
