+++ product/pro_edit_check.php
@@ -27,15 +27,16 @@ else
 
 <?php
 
-$pro_code = $_POST['code'];
-$pro_name = $_POST['name'];
-$pro_price = $_POST['price'];
-$pro_image_name_old = $_POST['image_name_old'];
-$pro_image = $_FILES['image'];
 
-$pro_code = htmlspecialchars($pro_code, ENT_QUOTES, 'UTF-8');
-$pro_name = htmlspecialchars($pro_name, ENT_QUOTES, 'UTF-8');
-$pro_price = htmlspecialchars($pro_price, ENT_QUOTES, 'UTF-8');
+require_once('../common/common.php');
+
+$post = sanitize($_POST);
+
+$pro_code = $post['code'];
+$pro_name = $post['name'];
+$pro_price = $post['price'];
+$pro_image_name_old = $post['image_name_old'];
+$pro_image = $_FILES['image'];
 
 if($pro_name=='')
 {
