+++ staff/staff_add_done.php
@@ -30,12 +30,11 @@ else
 try
 {
 
-$staff_name = $_POST['name'];
-$staff_pass = $_POST['pass'];
+require_once('../common/common.php');
 
-// 入力データに安全対策
-$staff_name = htmlspecialchars($staff_name, ENT_QUOTES, 'UTF-8');
-$staff_pass = htmlspecialchars($staff_pass, ENT_QUOTES, 'UTF-8');
+$post = sanitize($_POST);
+$staff_name = $post['name'];
+$staff_pass = $post['pass'];
 
 // データベースに接続
 $dsn = 'mysql:dbname=shop;host=localhost;charset=utf8';
