+++ staff/staff_edit_check.php
@@ -27,14 +27,22 @@ else
 
 <?php
 
-$staff_code = $_POST['code'];
-$staff_name = $_POST['name'];
-$staff_pass = $_POST['pass'];
-$staff_pass2 = $_POST['pass2'];
+require_once('../common/common.php');
 
-$staff_name = htmlspecialchars($staff_name, ENT_QUOTES, 'UTF-8');
-$staff_pass = htmlspecialchars($staff_pass, ENT_QUOTES, 'UTF-8');
-$staff_pass2 = htmlspecialchars($staff_pass2, ENT_QUOTES, 'UTF-8');
+$post = sanitize($_POST);
+$staff_code = $post['code'];
+$staff_name = $post['name'];
+$staff_pass = $post['pass'];
+$staff_pass2 = $post['pass2'];
+
+// $staff_code = $_POST['code'];
+// $staff_name = $_POST['name'];
+// $staff_pass = $_POST['pass'];
+// $staff_pass2 = $_POST['pass2'];
+
+// $staff_name = htmlspecialchars($staff_name, ENT_QUOTES, 'UTF-8');
+// $staff_pass = htmlspecialchars($staff_pass, ENT_QUOTES, 'UTF-8');
+// $staff_pass2 = htmlspecialchars($staff_pass2, ENT_QUOTES, 'UTF-8');
 
 if($staff_name=='')
 {
