+++ staff/staff_edit_done.php
@@ -30,12 +30,12 @@ else
 try
 {
 
-$staff_code = $_POST['code'];
-$staff_name = $_POST['name'];
-$staff_pass = $_POST['pass'];
+require_once('../common/common.php');
 
-$staff_name = htmlspecialchars($staff_name, ENT_QUOTES, 'UTF-8');
-$staff_pass = htmlspecialchars($staff_pass, ENT_QUOTES, 'UTF-8');
+$post = sanitize($_POST);
+$staff_code = $post['code'];
+$staff_name = $post['name'];
+$staff_pass = $post['pass'];
 
 $dsn = 'mysql:dbname=shop;host=localhost;charset=utf8';
 $user = 'root';
